+++ examples/ecca.py
@@ -106,7 +106,7 @@ plt.tight_layout()
 # -----------------
 # The full ERP CCA (eCCA) pipeline is implemented as a scikit-learn compatible class in PyntBCI in
 # `pyntbci.classifiers.eCCA`. All it needs are the lags if a circular shifted code is used (not used here) in `lags`,
-# the sampling frequency `fs`, and the the duration of one period of a code as `cycle_size`.
+# the sampling frequency `fs`, and the duration of one period of a code as `cycle_size`.
 #
 # When calling `eCCA.fit(X, y)` with training data `X` and labels `y`, the template responses are learned as well as the
 # spatial filters `eCCA.w_`.
@@ -126,7 +126,7 @@ ax.set_title("Spatial filter")
 # Cross-validation
 # ----------------
 # To perform decoding, one can call `eCCA.fit(X_trn, y_trn)` on training data `X_trn` and labels `y_trn` and
-# `eCCA.predict(X_tst)` on testing data `X_tst`. In this section, a chronological cross-validation is setup to evaluate
+# `eCCA.predict(X_tst)` on testing data `X_tst`. In this section, a chronological cross-validation is set up to evaluate
 # the performance of eCCA.
 
 trialtime = 4.2  # limit trials to a certain duration in seconds
@@ -149,7 +149,7 @@ for i_fold in range(n_folds):
     ecca.fit(X_trn, y_trn)
 
     # Apply template-matching classifier
-    yh_tst = ecca.predict(X_tst)[:, 0]  # select component
+    yh_tst = ecca.predict(X_tst)
 
     # Compute accuracy
     accuracy[i_fold] = np.mean(yh_tst == y_tst)
@@ -206,7 +206,7 @@ for i_fold in range(n_folds):
         ecca.fit(X_trn[:train_trials[i_trial], :, :], y_trn[:train_trials[i_trial]])
 
         # Apply classifier
-        yh_tst = ecca.predict(X_tst)[:, 0]  # select component
+        yh_tst = ecca.predict(X_tst)
 
         # Compute accuracy
         accuracy[i_fold, i_trial] = np.mean(yh_tst == y_tst)
@@ -261,7 +261,7 @@ for i_fold in range(n_folds):
     # Loop segments
     for i_segment in range(n_segments):
         # Apply classifier
-        yh_tst = ecca.predict(X_tst[:, :, :int(fs * segments[i_segment])])[:, 0]  # select component
+        yh_tst = ecca.predict(X_tst[:, :, :int(fs * segments[i_segment])])
 
         # Compute accuracy
         accuracy[i_fold, i_segment] = np.mean(yh_tst == y_tst)
@@ -329,7 +329,7 @@ for i_subject in range(n_subjects):
         ecca.fit(X_trn, y_trn)
 
         # Apply classifier
-        yh_tst = ecca.predict(X_tst)[:, 0]  # select component
+        yh_tst = ecca.predict(X_tst)
 
         # Compute accuracy
         accuracy[i_subject, i_fold] = np.mean(yh_tst == y_tst)