+++ pipelines/thielen2021_ecca.py
@@ -153,7 +153,7 @@ for i_fold in range(n_folds):
     ecca.fit(X_trn, y_trn)
 
     # Apply template-matching classifier
-    yh_tst = ecca.predict(X_tst)[:, 0]  # select component
+    yh_tst = ecca.predict(X_tst)
 
     # Compute accuracy
     accuracy[i_fold] = np.mean(yh_tst == y_tst)
@@ -210,7 +210,7 @@ for i_fold in range(n_folds):
         ecca.fit(X_trn[:train_trials[i_trial], :, :], y_trn[:train_trials[i_trial]])
 
         # Apply classifier
-        yh_tst = ecca.predict(X_tst)[:, 0]  # select component
+        yh_tst = ecca.predict(X_tst)
 
         # Compute accuracy
         accuracy[i_fold, i_trial] = np.mean(yh_tst == y_tst)
@@ -265,7 +265,7 @@ for i_fold in range(n_folds):
     # Loop segments
     for i_segment in range(n_segments):
         # Apply classifier
-        yh_tst = ecca.predict(X_tst[:, :, :int(fs * segments[i_segment])])[:, 0]  # select component
+        yh_tst = ecca.predict(X_tst[:, :, :int(fs * segments[i_segment])])
 
         # Compute accuracy
         accuracy[i_fold, i_segment] = np.mean(yh_tst == y_tst)
@@ -334,7 +334,7 @@ for i_subject in range(n_subjects):
         ecca.fit(X_trn, y_trn)
 
         # Apply classifier
-        yh_tst = ecca.predict(X_tst)[:, 0]  # select component
+        yh_tst = ecca.predict(X_tst)
 
         # Compute accuracy
         accuracy[i_subject, i_fold] = np.mean(yh_tst == y_tst)