+++ pipelines/thielen2021_rcca.py
@@ -231,7 +231,7 @@ for i_fold in range(n_folds):
     rcca.fit(X_trn, y_trn)
 
     # Apply template-matching classifier
-    yh_tst = rcca.predict(X_tst)[:, 0]  # select component
+    yh_tst = rcca.predict(X_tst)
 
     # Compute accuracy
     accuracy[i_fold] = np.mean(yh_tst == y_tst)
@@ -287,7 +287,7 @@ for i_fold in range(n_folds):
         rcca.fit(X_trn[:train_trials[i_trial], :, :], y_trn[:train_trials[i_trial]])
 
         # Apply classifier
-        yh_tst = rcca.predict(X_tst)[:, 0]  # select component
+        yh_tst = rcca.predict(X_tst)
 
         # Compute accuracy
         accuracy[i_fold, i_trial] = np.mean(yh_tst == y_tst)
@@ -342,7 +342,7 @@ for i_fold in range(n_folds):
     # Loop segments
     for i_segment in range(n_segments):
         # Apply classifier
-        yh_tst = rcca.predict(X_tst[:, :, :int(fs * segments[i_segment])])[:, 0]  # select component
+        yh_tst = rcca.predict(X_tst[:, :, :int(fs * segments[i_segment])])
 
         # Compute accuracy
         accuracy[i_fold, i_segment] = np.mean(yh_tst == y_tst)
@@ -411,7 +411,7 @@ for i_subject in range(n_subjects):
         rcca.fit(X_trn, y_trn)
 
         # Apply classifier
-        yh_tst = rcca.predict(X_tst)[:, 0]  # select component
+        yh_tst = rcca.predict(X_tst)
 
         # Compute accuracy
         accuracy[i_subject, i_fold] = np.mean(yh_tst == y_tst)